pyproject.toml
[build-system]
requires = ["setuptools >= 65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "franklin-cli"
version = "0.25.108"
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Education",
  "Topic :: Software Development :: Build Tools",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
requires-python = ">= 3.9" # it is actually ">=10" but if I add that conda build frails?...
dependencies = [
    'requests>=2.32.3',
    'colorama>=0.4.6',
    'readchar>=4.0.5',
    'click',
    'click-plugins',
    'psutil',
    'jupyter',
    'packaging',
    'pyyaml',
    'pyperclip',
    'selenium',
    'webdriver-manager',
    'rapidfuzz',
    'pycryptodome',
]
authors = [
  {name = "Kasper Munch", email = "kaspermunch@birc.au.dk"},
]
maintainers = [
  {name = "Kasper Munch", email = "kaspermunch@birc.au.dk"},
]
description = "Terminal client for Jupyter exercises"
keywords = ["teaching", "data", "jupyter", "docker", "git"]
readme = "README.md"

[project.urls]
Homepage = "https://munch-group.org/franklin"
Documentation = "https://munch-group.org/franklin"
Repository = "https://github.com/munch-group/franklin.git"
Issues = "https://github.com/munch-group/franklin/issues"

[project.scripts]
franklin = "franklin_cli:franklin"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
franklin_cli = [
  "data/*.yaml", 
  "data/gitui/*", 
  "data/templates/exercise/.devcontainer/devcontainer.json",
  "data/templates/exercise/.git",
  "data/templates/exercise/.gitattributes",
  "data/templates/exercise/.gitignore",
  "data/templates/exercise/.gitlab-ci.yml",
  "data/templates/exercise/.pixi.py",
  "data/templates/exercise/.test-notebook.sh",
  "data/templates/exercise/.vscode/settings.json",
  "data/templates/exercise/Dockerfile",
  "data/templates/exercise/exercise.ipynb",
  "data/templates/exercise/install_pixi.sh",
  "data/templates/exercise/pixi.toml",
  "data/templates/exercise/README.md",
  "data/templates/exercise/magic.py",
  ]
  
[tool.pixi.workspace]
channels = ["conda-forge", "munch-group"]
platforms = ["osx-arm64", "win-64", "osx-64", "linux-64"]

[tool.pixi.tasks.install-dev]
#cmd = "pip install --no-build-isolation --no-deps --disable-pip-version-check -e ."
cmd = "pip install -e ."
description = "Install the package in development mode"


# setup-kernel = "python -m ipykernel install --user --name franklin --display-name 'Franklin Python'"
# quarto-render = { cmd = "quarto render", env = { QUARTO_PYTHON = ".pixi/envs/default/bin/python" } }

[tool.pixi.tasks.bump-version]
args = [
  { "arg" = "kind", "default" = "patch" },  # can be patch, minor, major
  { "arg" = "flag", "default" = "" },     # can be --pre, --release
]
cmd = "./.scripts/bump_version.py --{{ kind }} {{ flag }} &&  git add -u &&  git commit -m '{{ kind }} {{ flag }} version bump' && git push" 
description = "Bump version patch and build/release conda packages"

[tool.pixi.tasks.bump-changelog]
cmd = "./.scripts/bump_changelog.py --yes"
description = "Bump version patch and build/release conda packages"

[tool.pixi.tasks.acp]
args = [
  { "arg" = "message", "default" = "update" },  # can be --patch, --minor, --major
]
cmd = "git add -u ; git commit -m '{{ message }}' ; git push"
description = "git add commit push of tracked files with optional message"

[tool.pixi.tasks.github-release]
cmd = "./.scripts/release-tag.sh"
description = "Release latest version on Github"


# devel feature
[tool.pixi.feature.devel.target.osx-arm64.dependencies]
setuptools = ">=65.0"
wheel = "*"
anthropic = "*"
gitpython = "*"
rpds-py = "*"

[tool.pixi.feature.devel.dependencies]
rpds-py = ">=0.27.0,<0.28"
jsonschema = ">=4.25.0,<5"


[tool.pixi.environments]
devel = ["devel"]

[tool.pixi.dependencies]
rpds-py = ">=0.27.0,<0.28"
  
#franklin = "0.25.8.*"


